LF-6017: potential integer overflow eliminated by casting to u64
Cast to u64, so that multiplications and additions will be done
in 64-bit arithmetic (and the results will also be 64-bit)
and the possibility of integer overflow is eliminated.

Fix coverity issue: Coverity Issues: 21652353 Unintentional integer overflow
http://coverity1.nxp.com/query/defects.htm?projectId=10289&cid=21652353

+++ drivers/clk/imx/clk-pll14xx.c
@@ -122,7 +122,7 @@ static long pll14xx_calc_rate(struct clk_pll14xx *pll, int mdiv, int pdiv,
 	}
 
 	/* fvco = (m * 65536 + k) * Fin / (p * 65536) */
-	fvco *= (mdiv * 65536 + kdiv);
+	fvco *= ((u64)mdiv * 65536 + (u64)kdiv);
 	pdiv *= 65536;
 
 	do_div(fvco, pdiv << sdiv);
